+++ diffsims/libraries/diffraction_library.py
@@ -20,6 +20,8 @@ import pickle
 
 import numpy as np
 
+from diffsims.generators.diffraction_generator import DiffractionGenerator
+from diffsims.utils._deprecated import deprecated
 
 __all__ = [
     "DiffractionLibrary",
@@ -114,6 +116,12 @@ class DiffractionLibrary(dict):
 
     """
 
+    @deprecated(
+        since="0.6.0",
+        alternative="diffsims.generators.simulation_generator.SimulationGenerator",
+        alternative_is_function=False,
+        removal="0.8.0",
+    )
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.identifiers = None
@@ -127,11 +135,11 @@ class DiffractionLibrary(dict):
 
         Parameters
         ----------
-        phase : str
+        phase : str or int
             Key for the phase of interest. If unspecified the choice is random.
         angle : tuple
             The orientation of interest as a tuple of Euler angles following the
-            Bunge convention [z, x, z] in degrees. If unspecified the choise is
+            Bunge convention [z, x, z] in degrees. If unspecified the choice is
             random (the first hit).
 
         Returns